Changed dropdown links, onchange function names
items.html: Changed the dropdown links to href #tableName i.e. #MSE. This redirects the page to a hash with the table. gatherData() does not need to be called anymore which is explained in my updated items.js description.

addItems.html ,makeRequest.html: changed the id's and onchange function calls to their new names in their respective js script files. Also changed dropdown links to href items.html#tableName, so the page will first redirect to items.html and then the hash. items.js is now commented out , and no longer linked to any of the pages except for items.html. This is because of the even listeners and DOMContentLoaded functions I have in items.js which will be explained in those updates.

home.html: this is TJ's home page html design. I just added the current sidebar html, the new dropdown links and changed the file name to home.html.

checkInOut.html: same changes as the other form html's, except I removed the call to the checkout(action) function onclick of the check out or check in buttons, since the call is now made in the checkInOut.js file. Explanation is provided in that update description.