Changed Variable Names, linked items.js to items.html
 Changed Variable Names, linked items.js to items.html

Items.js: Provided new javascript between the gatherData() and drawResults() functions in lines 185-216. These functions are responsible for handling the hashes added in the href links for each dropdown table name. Comments are provided with descriptions in each method. DOMContentLoaded Event listener is added to refresh the page to load the correct hash. Because of this, the other files content is always emptied if the html file is linked to items.js, so that link has to be removed. Because of this, checkout() post function for routing is moved to checkInOut.js. Also modified line 248 so the header for the tables will always be capitalized. In the beginning of the gatherData() function, I also added an Object Map to display the html headers of each table correctly after the change in the hashes URL, as well as an if statement that ensures the table styling always uses the content id, and never the styling id of the previous page.

checkInoutJs, addItem.js, makeRequest.js: changed main object in the first line to itemsAndAttributes, and then changed all "category" names to "item" and subcategory names to "attributes" for better readability. The original variable names were with respect to the cables, hardware, peripherals categories which we removed. Also made minor changes for more professional front end display. function displaySelectedItem() at line 330 is to display what item users selected in the dropdown for the confirmation modal so users can double check which item they selected before they officially submit the form.

In addItems.js and makeRequest.js, I added a function handleSubmit() at the very bottom on line 328 that prevents the default form actions from occuring, which is resetting the page oncllick of the submit form button, that way the modals will still display. This default action occurs since there is only one submit button, opposed to two submit buttons for the check in/out form.

Please let me know if you would like me to elaborate further or have any specific questions