Finalized Success Modal, makeRequest.js and html, delete item
Changes discussed in group chat.

The specific changes in makeRequest.js: I changed the object key option and attribute names to be more user friendly and comprehensible, since make a request does not need to match the database variable or table names. In the handleSubmit function on line 274 of makeRequest is where I added the mailTo: route when the submit request button is clicked. All the routing functions had changes made with a .then(response => { for the client-end functions, and then I created the res.json() responses to send back on the server.js in the try and catch conditions for each.